move parseNumberValue to shared utils code
