+++ frontend/src/metabase/components/TokenField/utils.ts
@@ -1,13 +1,3 @@
-export function parseNumberValue(value: any): number | null {
-  const number = Number.parseFloat(value);
-
-  if (Number.isFinite(number)) {
-    return number;
-  } else {
-    return null;
-  }
-}
-
 export function parseStringValue(value: any): string | null {
   const trimmedValue = trim(value);
   if (trimmedValue === "") {
