+++ frontend/src/metabase/lib/number.unit.spec.ts
@@ -1,4 +1,4 @@
-import { isPositiveInteger } from "./number";
+import { isPositiveInteger, parseNumberValue } from "./number";
 
 describe("metabase/lib/number", () => {
   describe("isPositiveInteger", () => {
@@ -27,6 +27,33 @@ describe("metabase/lib/number", () => {
       expect(isPositiveInteger("123abc")).toBe(false);
       expect(isPositiveInteger("abc123")).toBe(false);
       expect(isPositiveInteger("abc")).toBe(false);
+    });
+  });
+
+  describe("parseNumberValue", () => {
+    it("should return null for non-number values", () => {
+      expect(parseNumberValue("")).toBeNull();
+      expect(parseNumberValue(" ")).toBeNull();
+      expect(parseNumberValue(" foo ")).toBeNull();
+      expect(parseNumberValue(null)).toBeNull();
+      expect(parseNumberValue("abc 123")).toBeNull();
+    });
+
+    it("should return null for numbers that are not finite", () => {
+      expect(parseNumberValue("Infinity")).toBeNull();
+      expect(parseNumberValue(Infinity)).toBeNull();
+      expect(parseNumberValue("-Infinity")).toBeNull();
+      expect(parseNumberValue(NaN)).toBeNull();
+      expect(parseNumberValue("NaN")).toBeNull();
+    });
+
+    it("should return a value parsed as a float", () => {
+      expect(parseNumberValue("123")).toBe(123);
+      expect(parseNumberValue("123px")).toBe(123);
+      expect(parseNumberValue("123.456")).toBe(123.456);
+      expect(parseNumberValue("0")).toBe(0);
+      expect(parseNumberValue(0)).toBe(0);
+      expect(parseNumberValue(123)).toBe(123);
     });
   });
 });
